Security, responsiveness and features update - Now it is possible to reload language while the bot is running - Fixed issue that could let bot restart take up to 1 hour - Added more secure way to change guild and user data (it wasn't safe, the design originally allowed that) - Added guild default poll emojis - Updated D++ lib - Updated language files - Fixed possible kill while uploading files to Discord - Fixed int32_t to int64_t (D++ update) - Moved multi emoji parsing to tools.hpp - Updated command version

+++ src/configurer.cpp
@@ -42,8 +42,8 @@ unsigned long long ConfigSetting::qgl(const std::string& key)
     return json[key].get<unsigned long long>();
 }
 
-int32_t ConfigSetting::qgi(const std::string& key)
+int64_t ConfigSetting::qgi(const std::string& key)
 {
     std::lock_guard<std::mutex> luck(mu);
-    return json[key].get<int32_t>();
+    return json[key].get<int64_t>();
 }
